Modified watchmakers for clarity. Functions are now seperated into different functions.
